feat: changed the layout of the main page and added condition for the navbar
