+++ src/components/Navbar.tsx
@@ -1,10 +1,25 @@
-import { Menu, Sparkles, X } from "lucide-react";
-import { useState } from "react";
+import { Bell, Menu, Sparkles, X } from "lucide-react";
+import { useEffect, useState } from "react";
+import Cookies from "js-cookie"; // Install this package if not already installed
 
 interface NavbarProps {}
 
 const NavBar: React.FC<NavbarProps> = () => {
   const [mobileMenuOpen, setMobileMenuOpen] = useState(false);
+  const [isLoggedIn, setIsLoggedIn] = useState(false);
+  const [notifications] = useState(2); // Example notification count
+
+  useEffect(() => {
+    const authToken = Cookies.get("auth_token"); // Replace "auth_token" with the actual cookie name
+    setIsLoggedIn(!!authToken);
+  }, []);
+
+  // Mock user data
+  const user = {
+    name: "Alex Johnson",
+    avatar: "/api/placeholder/32/32",
+    plan: "Pro",
+  };
 
   return (
     <nav className="fixed top-0 left-0 right-0 z-50 backdrop-blur-xl bg-slate-900/80 border-b border-white/10">
@@ -37,12 +52,40 @@ const NavBar: React.FC<NavbarProps> = () => {
             >
               Pricing
             </a>
-            <button
-              onClick={() => (window.location.href = "/auth/register")}
-              className="px-4 py-2 bg-gradient-to-r from-purple-500 to-pink-500 rounded-lg font-medium hover:opacity-90 transition-opacity"
-            >
-              Register
-            </button>
+
+            {isLoggedIn ? (
+              <div className="flex items-center gap-4">
+                <div className="relative">
+                  <Bell className="w-6 h-6 text-gray-300 cursor-pointer hover:text-white transition-colors" />
+                  {notifications > 0 && (
+                    <span className="absolute -top-1 -right-1 bg-red-500 text-white text-xs rounded-full w-4 h-4 flex items-center justify-center">
+                      {notifications}
+                    </span>
+                  )}
+                </div>
+
+                <div className="flex items-center gap-3 cursor-pointer pl-4 border-l border-gray-700">
+                  <img
+                    src={user.avatar}
+                    alt="User avatar"
+                    className="w-8 h-8 rounded-full border border-purple-400"
+                  />
+                  <div>
+                    <p className="text-sm font-medium text-white">
+                      {user.name}
+                    </p>
+                    <p className="text-xs text-purple-400">{user.plan}</p>
+                  </div>
+                </div>
+              </div>
+            ) : (
+              <button
+                onClick={() => (window.location.href = "/auth/register")}
+                className="px-4 py-2 bg-gradient-to-r from-purple-500 to-pink-500 rounded-lg font-medium hover:opacity-90 transition-opacity"
+              >
+                Register
+              </button>
+            )}
           </div>
 
           {/* Mobile Menu Button */}
@@ -81,15 +124,29 @@ const NavBar: React.FC<NavbarProps> = () => {
             >
               Pricing
             </a>
-            <button
-              onClick={() => {
-                setMobileMenuOpen(false);
-                window.location.href = "/image-gen";
-              }}
-              className="w-full px-4 py-2 bg-gradient-to-r from-purple-500 to-pink-500 rounded-lg font-medium hover:opacity-90 transition-opacity"
-            >
-              Register
-            </button>
+            {isLoggedIn ? (
+              <div className="flex items-center gap-3">
+                <img
+                  src={user.avatar}
+                  alt="User avatar"
+                  className="w-8 h-8 rounded-full border border-purple-400"
+                />
+                <div>
+                  <p className="text-sm font-medium text-white">{user.name}</p>
+                  <p className="text-xs text-purple-400">{user.plan}</p>
+                </div>
+              </div>
+            ) : (
+              <button
+                onClick={() => {
+                  setMobileMenuOpen(false);
+                  window.location.href = "/auth/register";
+                }}
+                className="w-full px-4 py-2 bg-gradient-to-r from-purple-500 to-pink-500 rounded-lg font-medium hover:opacity-90 transition-opacity"
+              >
+                Register
+              </button>
+            )}
           </div>
         </div>
       )}
